+++ evkmimxrt685/driver_examples/powerquad/fir_fast/powerquad_fir_fast.c
@@ -462,7 +462,7 @@ static void PQ_FIRFloatExample(void)
 
     for (i = 0; i < EXAMPLE_FIR_DATA_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_firOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_firOutputRef[i] - output[i])) < 0.00001);
     }
 
     /*
@@ -508,7 +508,7 @@ static void PQ_FIRFloatExample(void)
 
     for (i = 0; i < EXAMPLE_FIR_DATA_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_firOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_firOutputRef[i] - output[i])) < 0.00001);
     }
 }
 
@@ -756,7 +756,7 @@ static void PQ_ConvolutionFloatExample(void)
 
     for (i = 0; i < EXAMPLE_CONV_RESULT_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_convOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_convOutputRef[i] - output[i])) < 0.00001);
     }
 
     /*
@@ -802,7 +802,7 @@ static void PQ_ConvolutionFloatExample(void)
 
     for (i = 0; i < EXAMPLE_CONV_RESULT_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_convOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_convOutputRef[i] - output[i])) < 0.00001);
     }
 }
 
@@ -1050,7 +1050,7 @@ static void PQ_CorrelationFloatExample(void)
 
     for (i = 0; i < EXAMPLE_CORR_RESULT_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_corrOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_corrOutputRef[i] - output[i])) < 0.00001);
     }
 
     /*
@@ -1096,6 +1096,6 @@ static void PQ_CorrelationFloatExample(void)
 
     for (i = 0; i < EXAMPLE_CORR_RESULT_LEN; i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(s_corrOutputRef[i] - output[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(s_corrOutputRef[i] - output[i])) < 0.00001);
     }
 }
