+++ mimxrt685audevk/driver_examples/powerquad/benchmark_sw_filter/powerquad_benchmark_sw_filter.c
@@ -281,7 +281,7 @@ static void arm_fir_f32Test(void)
 
     for (i = 0; i < ARRAY_SIZE(FIRRef); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(FIRRef[i] - FIRResult[i]) < 0.0001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(FIRRef[i] - FIRResult[i])) < 0.0001);
     }
 }
 
@@ -403,7 +403,7 @@ static void arm_conv_f32Test(void)
 
     for (uint32_t i = 0; i < ARRAY_SIZE(ref); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(ref[i] - result[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(ref[i] - result[i])) < 0.00001);
     }
 }
 
@@ -525,6 +525,6 @@ static void arm_correlate_f32Test(void)
 
     for (uint32_t i = 0; i < ARRAY_SIZE(ref); i++)
     {
-        EXAMPLE_ASSERT_TRUE(fabs(ref[i] - result[i]) < 0.00001);
+        EXAMPLE_ASSERT_TRUE(fabs((double)(ref[i] - result[i])) < 0.00001);
     }
 }
